# docs/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "GrownupsVet API",
    "description" : "Contrato generado desde código para acceso, perfil, sesiones, mascotas, recuperación, personal con invitaciones y disponibilidad veterinaria. El correo requiere configuración SMTP. La URL de activación del portal contiene un placeholder configurable. Reservas, consulta automática de COMVEZCOL y plantillas Thymeleaf quedan fuera de este incremento.",
    "version" : "0.6.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/staff/{userId}/invitations" : {
      "post" : {
        "tags" : [ "Invitaciones de personal" ],
        "summary" : "Reenviar invitación de personal",
        "description" : "SUPER_ADMIN invita ADMINISTRATOR; ADMINISTRATOR invita VETERINARIAN. Solo cuentas pendientes o deshabilitadas sin contraseña. Invalida el enlace anterior. 202 acredita la cola persistente; no confirma entrega SMTP.",
        "operationId" : "resendStaffInvitation",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "401" : {
            "description" : "Sesión ausente, inválida o revocada",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "El rol actual no puede administrar esta cuenta",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Cuenta no encontrada",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "La cuenta ya tiene credenciales o está activa",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "429" : {
            "description" : "Máximo cinco invitaciones por cuenta en 24 horas y un minuto entre solicitudes",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "503" : {
            "description" : "Canal de invitaciones deshabilitado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "202" : {
            "description" : "Nueva invitación y tarea de correo persistidas",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffInvitationResponseDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Invitaciones de personal" ],
        "summary" : "Cancelar invitación de personal",
        "description" : "Revoca enlaces pendientes, elimina el secreto temporal de entrega y deja la cuenta sin credenciales deshabilitada.",
        "operationId" : "cancelStaffInvitation",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "401" : {
            "description" : "Sesión ausente, inválida o revocada",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "403" : {
            "description" : "El rol actual no puede administrar esta cuenta",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Cuenta no encontrada",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "La cuenta ya tiene credenciales o está activa",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "429" : {
            "description" : "Máximo cinco invitaciones por cuenta en 24 horas y un minuto entre solicitudes",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "503" : {
            "description" : "Canal de invitaciones deshabilitado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "204" : {
            "description" : "Invitación cancelada; repetir sobre la cuenta ya deshabilitada es idempotente"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/auth/account-activations" : {
      "post" : {
        "tags" : [ "Activación de cuenta de personal" ],
        "summary" : "Activar cuenta de personal mediante invitación",
        "description" : "Consume un enlace opaco de un uso, válido durante 48 horas. Solo activa cuentas pendientes sin credenciales. Aplica la política de contraseña del registro, con confirmación idéntica. No crea automáticamente una sesión.",
        "operationId" : "activateStaffAccount",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StaffAccountActivationRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Cuenta activada; puede iniciar sesión con la contraseña elegida"
          },
          "400" : {
            "description" : "Campos inválidos, confirmación distinta o enlace inválido, vencido o consumido",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "415" : {
            "description" : "El cuerpo debe ser application/json",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          },
          "503" : {
            "description" : "Activación por invitación deshabilitada",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiErrorResponseDTO" : {
        "type" : "object",
        "description" : "Respuesta de error RFC 9457, con código funcional y errores por campo.",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "Referencia URI al tipo de problema.",
            "example" : "about:blank"
          },
          "title" : {
            "type" : "string",
            "description" : "Resumen del problema.",
            "example" : "Solicitud incorrecta"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Estado HTTP de la respuesta.",
            "example" : 400,
            "maximum" : 599,
            "minimum" : 400
          },
          "detail" : {
            "type" : "string",
            "description" : "Explicación de lo ocurrido en esta solicitud.",
            "example" : "Revisa los campos indicados."
          },
          "instance" : {
            "type" : "string",
            "format" : "uri-reference",
            "description" : "Referencia URI opcional que identifica esta ocurrencia del problema.",
            "example" : "urn:uuid:7d667530-867b-4c78-a458-0904fb82d574"
          },
          "errorCode" : {
            "type" : "string",
            "description" : "Código funcional estable; el frontend no debe comparar los mensajes.",
            "example" : "INVALID_REQUEST"
          },
          "fieldErrors" : {
            "type" : "array",
            "description" : "Errores por campo. Es una lista vacía cuando el error no corresponde a campos específicos.",
            "items" : {
              "$ref" : "#/components/schemas/FieldValidationErrorResponseDTO"
            }
          }
        },
        "required" : [ "detail", "errorCode", "fieldErrors", "status", "title", "type" ]
      },
      "FieldValidationErrorResponseDTO" : {
        "type" : "object",
        "description" : "Error asociado a un campo de la solicitud. Nunca contiene el valor recibido.",
        "properties" : {
          "field" : {
            "type" : "string",
            "description" : "Nombre del campo en el contrato JSON.",
            "example" : "email"
          },
          "code" : {
            "type" : "string",
            "description" : "Código estable que identifica el error del campo.",
            "example" : "TYPE_MISMATCH"
          },
          "message" : {
            "type" : "string",
            "description" : "Explicación legible para la persona usuaria.",
            "example" : "Debe enviarse como texto."
          }
        },
        "required" : [ "code", "field", "message" ]
      },
...
      "StaffInvitationResponseDTO" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "invitationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "userId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Vencimiento del enlace de un uso, 48 horas desde su creación."
          },
          "deliveryStatus" : {
            "type" : "string",
            "description" : "Envío persistido y pendiente; no confirma recepción del correo.",
            "enum" : [ "PENDING" ]
          }
        },
        "required" : [ "deliveryStatus", "expiresAt", "invitationId", "userId" ]
      },
...
      "StaffAccountActivationRequestDTO" : {
        "type" : "object",
        "additionalProperties" : false,
        "properties" : {
          "token" : {
            "type" : "string",
            "maxLength" : 43,
            "minLength" : 43,
            "pattern" : "[A-Za-z0-9_-]{43}",
            "writeOnly" : true
          },
          "password" : {
            "type" : "string",
            "format" : "password",
            "maxLength" : 128,
            "minLength" : 15,
            "writeOnly" : true
          },
          "confirmPassword" : {
            "type" : "string",
            "format" : "password",
            "description" : "Debe coincidir exactamente con password, incluidos espacios y mayúsculas.",
            "maxLength" : 128,
            "minLength" : 15,
            "writeOnly" : true
          }
        },
        "required" : [ "confirmPassword", "password", "token" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "JWT RS256 supplied in the Authorization header. Tokens expire 24 hours after login.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}